# lib/openapi/spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Timesheet API",
    "version": "1.0.0",
    "description": "Complete REST API for the Timesheet workforce management system. Handles employee time tracking, scheduling, roster management, payroll and analytics.\n\n## Authentication\nUses cookie-based JWT authentication. Three separate auth layers:\n- Admin/user: login via /api/auth/login\n- Employee: login via /api/employee/login\n- Device: register via /api/device/register\n\n## Roles\n- super_admin: full system access\n- admin: manage employees and schedules\n- user: limited view-only dashboard access\n- employee: clock in/out and own timesheets\n\n## Base URL\nAll endpoints are prefixed with /api\n\n## Error Responses\nAll errors follow the format: { error: string, message?: string }\n\n## Cookie Auth\nAll protected endpoints require valid JWT cookie sent automatically via credentials: include on all requests."
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "message": {
            "type": "string",
            "example": "Invalid input data"
          },
          "details": {
            "type": "object"
          }
        },
        "required": ["error"]
      },
...
      "ObjectId": {
        "type": "string",
        "pattern": "^[0-9a-fA-F]{24}$",
        "example": "507f1f77bcf86cd799439011"
      },
      "DateTime": {
        "type": "string",
        "format": "date-time",
        "example": "2024-03-07T10:30:00.000Z"
      },
      "Date": {
        "type": "string",
        "format": "date",
        "example": "2024-03-07"
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "user", "super_admin"],
            "example": "user"
          },
          "location": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "rights": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["read_employees", "write_timesheets"]
          },
          "managedRoles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "createdAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/DateTime"
          }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "example": "john_doe"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "securePassword123"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "user", "super_admin"],
            "example": "user"
          },
          "location": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "rights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "managedRoles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        },
        "required": ["username", "password", "role"]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "minimum": 1,
            "example": 1
          },
          "limit": {
            "type": "number",
            "minimum": 1,
            "maximum": 100,
            "example": 20
          },
          "offset": {
            "type": "number",
            "minimum": 0,
            "example": 0
          },
          "total": {
            "type": "number",
            "example": 150
          }
        }
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users with pagination and search",
        "description": "Returns a paginated list of all admin and user accounts in the system. Supports search filtering by username and pagination controls. Includes user roles, permissions, assigned locations, and account status. Only accessible by admin and super_admin roles. User role can only see limited information about other users.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search term to filter users by username. Case-insensitive partial matching."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create new admin or user account",
        "description": "Creates a new admin or user account with specified role and permissions. Validates username uniqueness and password strength requirements. Assigns location access and role-based permissions. Only accessible by super_admin and admin roles. Automatically hashes passwords using bcrypt before storage. Optionally sends setup email for initial password configuration.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/timesheets/payroll": {
      "get": {
        "tags": ["Timesheets"],
        "summary": "Get payroll data with hours and pay calculations",
        "description": "Retrieves comprehensive payroll data including total hours, regular hours, overtime hours, and calculated gross pay for employees. Supports filtering by date range and employee for payroll processing. Includes hourly rates and pay calculations based on award classifications. Essential for payroll preparation and labor cost analysis.",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Date"
            }
          },
          {
            "name": "employeeId",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payroll data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "employeeId": {
                        "$ref": "#/components/schemas/ObjectId"
                      },
                      "employeeName": {
                        "type": "string",
                        "example": "John Doe"
                      },
                      "totalHours": {
                        "type": "number",
                        "example": 80.5
                      },
                      "regularHours": {
                        "type": "number",
                        "example": 80.0
                      },
                      "overtimeHours": {
                        "type": "number",
                        "example": 0.5
                      },
                      "grossPay": {
                        "type": "number",
                        "example": 2012.50
                      },
                      "hourlyRate": {
                        "type": "number",
                        "example": 25.00
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}